==> manejo-florestal-jamari/01_prep_data/part1-prep-data-traitsJamari.rmd <==
---
title: "traitsJamari"
author: "Fernando Lima"
date: "10/02/2022"
output:
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#Limpar o cache:
rm(list=ls())
#Pacotes para instalar caso necessário:
#Chamando pacotes
library(ade4)
library(clue)
library(picante)
library(here)
```

## Abrir Dados

```{r dados}
#traitsMammals <- read.csv(here("data","traits_202006d22.csv"), head=T,  row.names=1, sep = ";")
library(gsheet)
library(tidyverse)
traitsMammals <- as_tibble(gsheet2tbl("https://docs.google.com/spreadsheets/d/13pazl0Qvzim1kQmDtM59mCIsBoZfYppwNEq7xYtkYk0/edit?usp=sharing"))
attach(traitsMammals)
```

## Selecionar traits

```{r selecao}
tabQ <- traitsMammals$Peso
tabN <- traitsMammals[,c("Carnivoro","Insetivoro","Frugivoro","Herbivoro","Terrestre","Fossorial")] # 10,11,12,14,15,16
```

## Verificação

```{r}
tabN
```

```{r dataPrep}
ktab1 <- ktab.list.df(list(data.frame(tabN),data.frame(tabQ)))
distrait <- dist.ktab(ktab1, c("N", "Q"), c("scaledBYrange"))

disU <- ls_fit_ultrametric(distrait)
treeU <- hclust2phylog(hclust(disU, "average"))
tree<-as.phylo(treeU)
```

## Plots

```{r pressure, echo=FALSE}
plot(treeU)
#plot(tree)
```
